# src/web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FamilyJobsBoard.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5099/"
    }
  ],
  "paths": {
...
    "/api/good-behaviour-types": {
      "get": {
        "tags": [
          "Good behaviours"
        ],
        "summary": "List the good behaviour types adults can log.",
        "description": "Returns active types with the points each usually earns. Any signed-in family member may read this list; the amount awarded for a logged behaviour can differ.",
        "operationId": "ListGoodBehaviourTypes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodBehaviourTypesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Good behaviours"
        ],
        "summary": "Create a good behaviour type.",
        "description": "Adult-only. The type is immediately available to log for any active child.",
        "operationId": "CreateGoodBehaviourType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGoodBehaviourTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodBehaviourTypeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/good-behaviour-types/{id}": {
      "put": {
        "tags": [
          "Good behaviours"
        ],
        "summary": "Edit a good behaviour type.",
        "description": "Adult-only. Previously logged behaviours keep the name, description, and points they were logged with. Returns 409 for a deleted type.",
        "operationId": "UpdateGoodBehaviourType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveGoodBehaviourTypeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoodBehaviourTypeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Good behaviours"
        ],
        "summary": "Delete a good behaviour type.",
        "description": "Adult-only soft delete: the type can no longer be logged, and history is unchanged. Repeating the request succeeds.",
        "operationId": "DeleteGoodBehaviourType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/good-behaviours": {
      "post": {
        "tags": [
          "Good behaviours"
        ],
        "summary": "Log a good behaviour for a child and award its points.",
        "description": "Adult-only. Creates the behaviour and its points ledger entry in one transaction. Points default to the type's points. Repeating a request ID with the same details returns the original result without a second award; reusing it for different details returns 409.",
        "operationId": "LogGoodBehaviour",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogGoodBehaviourRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogGoodBehaviourResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogGoodBehaviourResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "GoodBehaviourResponse": {
        "required": [
          "id",
          "typeId",
          "typeName",
          "typeDescription",
          "childId",
          "loggedByMemberId",
          "points",
          "loggedAtUtc"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "typeName": {
            "type": "string"
          },
          "typeDescription": {
            "type": "string"
          },
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "loggedByMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "loggedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GoodBehaviourTypeResponse": {
        "required": [
          "id",
          "name",
          "description",
          "points"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "points": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "GoodBehaviourTypesResponse": {
        "required": [
          "types"
        ],
        "type": "object",
        "properties": {
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GoodBehaviourTypeResponse"
            }
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "LogGoodBehaviourRequest": {
        "required": [
          "requestId",
          "typeId",
          "childId",
          "points"
        ],
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "childId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "LogGoodBehaviourResponse": {
        "required": [
          "behaviour",
          "pointsBalance"
        ],
        "type": "object",
        "properties": {
          "behaviour": {
            "$ref": "#/components/schemas/GoodBehaviourResponse"
          },
          "pointsBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveGoodBehaviourTypeRequest": {
        "required": [
          "name",
          "description",
          "points"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "points": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Short-lived access token returned by the household sign-in flow.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}